requirements.txt
Django==2.2.18
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.11.0
django-rest-swagger
drf-yasg
djangorestframework_simplejwt==4.6.0
